Remove binding default objects (0 handle) to OpenGL targets
This allows to avoid an OpenGL warning (NVidia) when a texture unit is activated
via glActiveTexture() but some texture is updated and glBindTexture(0) is called
in the end.

For some reason this produces another one during glClear() with a particular FBO
bound to GL_DRAW_FRAMEBUFFER target:
> Program/shader state usage warning: A fragment program/shader is required
> to correctly render to an integer framebuffer.

Ignore this one, glClear() is program independent.